Add 'Support Keyboard' feature and integrate module
This commit introduces a new 'Support Keyboard' feature,
encapsulated within a new `:ime:kasahorow` module.

Key changes include:
- Added `SupportKeyboardView` and `SupportKeyboardController`
to manage the UI and logic for launching the support keyboard website.
- The `SupportKeyboardView` is a custom `FrameLayout` that displays a
title and handles click events to launch the support page.
- The `SupportKeyboardController` constructs the URL for
the support page based on the current keyboard and source screen,
and launches an intent to open it in a browser.
- Integrated the `SupportKeyboardView` as a header in the add-on
browser screen (`AbstractAddOnsBrowserFragment`).
This allows users to access the support page directly from the add-on settings.